refactor : AllocateTaskRequest 경로 변경
