+++ back/src/main/java/shop/tukoreamyway/back/domain/task/dto/AllocateTaskRequest.java
@@ -1,4 +1,4 @@
-package shop.tukoreamyway.back.domain.task.command.application;
+package shop.tukoreamyway.back.domain.task.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
